# chore: removed unnecessary array indexing proofs