--- a/Std/CodeAction/Deprecated.lean
+++ b/Std/CodeAction/Deprecated.lean
@@ -32,7 +32,7 @@ def deprecatedCodeActionProvider : CodeActionProvider := fun params snap => do
   for diag in snap.interactiveDiags do
     if let some #[.deprecated] := diag.tags? then
       if h : _ then
-        msgs := msgs.push (snap.cmdState.messages.msgs[i]'h)
+        msgs := msgs.push (snap.cmdState.messages.msgs[i])
     i := i + 1
   if msgs.isEmpty then return #[]
   let start := doc.meta.text.lspPosToUtf8Pos params.range.start