Inference of array indexing proofs has improved. This PR removes those that no longer need to be given explicitly.

I *think* this is an improvement but would like a second opinion.