@digama0 2024-03-19
Hm, too bad we don't have !bench here. Did you notice any performance degradation with this?