@semorrison 2024-03-24
> Hm, too bad we don't have !bench here. Did you notice any performance degradation with this?

Before:
```
% lake build && hyperfine "rm -f .lake/build/lib/Std/Data/Array/Lemmas.olean && lake build Std.Data.Array.Lemmas"
Benchmark 1: rm -f .lake/build/lib/Std/Data/Array/Lemmas.olean && lake build Std.Data.Array.Lemmas
  Time (mean ± σ):      3.555 s ±  0.017 s    [User: 3.305 s, System: 0.395 s]
  Range (min … max):    3.528 s …  3.582 s    10 runs
```
After
```
% lake build && hyperfine "rm -f .lake/build/lib/Std/Data/Array/Lemmas.olean && lake build Std.Data.Array.Lemmas"
Benchmark 1: rm -f .lake/build/lib/Std/Data/Array/Lemmas.olean && lake build Std.Data.Array.Lemmas
  Time (mean ± σ):      3.760 s ±  0.029 s    [User: 3.497 s, System: 0.398 s]
  Range (min … max):    3.712 s …  3.805 s    10 runs
```

So there is a measurable slow-down. I'm inclined to say it's worth it, still?